Basic recipe for Miniforge
Provides a Python script, `make.py`, which creates a temporary directory
for the build to proceed and fills out some information used by the
constructor to specify the name, version, and `python` used. Construct
the version using the UTC associated with the commit used to generate
the build. Ensures the resulting installer is placed in `out/` and the
temporary directory deleted. Ensure that `git` ignores the `out/`
directory.