Provide option to generate hashes for installers
Adds a flag to `build.py` to allow a hash or multiple hashes to be
specified and computed for the installers. Anything supported by
`hashlib`/`openssl` is allowed. The hashes are written out to files that
share the same name as the installer, but with the checksum appended as
an extension.