Add stubs for Unix and Windows check scripts
Provides bash/batch scripts for testing the installer and runs them as
part of `check.py`. The scripts currently don't do anything yet. However
they will be used for running the installer and inspecting the installed
result to make sure it passes some criteria. Scripts are run in a
temporary directory where the installer is also placed. The temporary
directory is cleaned after the tests complete.