Drop `defaults` from `channels`
Builds the installer using only the `conda-forge` channel without
`defaults` (based off of `conda` 4.3). As all needed packages are now
available in `conda-forge`, it makes it easy to use `conda-forge` as the
source of truth for all dependencies. Though users are still free to use
other channels as they like after running the installer.