Run `distclean` after CIs complete build and test
In an effort to test the `distclean` script and clean up our installers,
run `distclean` after `build`. The cleanliness is not necessary right
now as the CI image instances are tossed afterwards. However cleaning
proactively will be important when we start caching in some cases to
make sure we don't build up too much junk.