Consolidate CircleCI steps into one script
Instead of having multiple CircleCI `run`/`command` lines, just have one
for each step that calls our script. Then run the rest of the steps from
the script. This should be faster as only one container is initiated.
Also should be easier to maintain as all the code is in one file.