Make sure to run `distclean` before CI builds
In order to prepare for caching, make sure that we are clearing the
results from any previous build before starting a new build. This way we
are not allowing the cache to accumulate between builds. Only allowing
the cache to include whatever we built from the current build where we
didn't cleanup afterwards.