Try caching the installers on CIs for release
In order to guarantee that all installers built ok and are ready to
release before actually tagging and releasing, we need to block tagging
until we know all CIs have passed. We have decided to do this by having
Heroku create the tag using a simple Python script as Heroku can block
until all CIs pass. The tag will trigger CIs to build again. However we
don't want to rebuild the installers a second time (especially if
something changes between builds). Instead we would rather release the
installers we built on `master`. The issue becomes how to pass the
installers between `master` and the tag. We accomplish this by caching
with CI, which is what this attempts.