Fetch remote tag before tagging
If a remote is provided, fetch the tag we are about to make from it
before creating it. This way we can be sure we don't tag the same thing
again. Unless of course we want to forcibly overwrite this tag, in which
case this tag will be overwritten immediately after it is fetched.