Provide option to publish releases to GitHub
In order to upload artifacts, we need a preexisting release on GitHub.
That release can be a prerelease or not. So provide an option in the
`release.py` script to publish a prerelease or a final release. In
order to make this work smoothly with pushing and handle changing the
release type, allow the remote to be specified independently of whether
to push or not.

[skip appveyor]
[appveyor skip]