Specify the `Content-Length` header for uploads
In Python 3.6, it appears there was a change that causes uploads to be
chunked by default if the `Content-Length` is not specified. AFAIK this
behavior is not supported by GitHub releases. It also results in the
connection being closed and reopened while we still have it open
triggering `ConnectionResetError`s. To fix this, we send the
`Content-Length` in the header after measuring it from the memory mapped
file.

ref: https://bugs.python.org/issue12319

[ci skip]