Only publish files including the current version
Adds a check to make sure that each filename includes the current
version before uploading them. If any filename does not include the
current version, skip over that file. This should fix a bug that we have
seen where old versions of the installer lingering in the cache were
being uploaded to a newer version.

[ci skip]