Instead of `prerelease` use `draft` with GitHub
Using `prerelease` with our partially completed releases is a bit of a
misnomer. Not to mention while our release is still in progress
`prerelease` allows others to download installers before the release
process is complete. In contrast a `draft` release remains hidden until
it is no longer a `draft`. Though it is still available for developers
to explore should the release break or if someone is curious about the
status of a release. As such we switch from using `prerelease` to using
`draft` for a smoother release experience.