More informative upload process
Now checks to see if a file is already uploaded and skips it if it is
(sharing that information with the user). Also prints out all files
skipped due to version mismatch, which should provide more insight into
the cache's contents.

In addition to sharing a bit more information and avoid failing when a
file is already up. Check after uploading a file that it actually shows
up in the assets for the release. If the file is not present, error out
the build. This should help catch cases like those seen on AppVeyor.

[ci skip]
[skip ci]