Ensure the CIs' cache directory is always present
This step ensures that the `out` cache directory always exists on CIs
even if `distclean.py` was run. Since we don't know what the behavior of
the CIs are when the directory is missing, it makes more sense to have
the cache directory present, but ensure it is empty of all content when
it should be.

Further as AppVeyor issued warnings about the directory being missing,
this should fix those warnings. Also may help with AppVeyor's lingering
cache issues.