Test the installer spec matches exactly
Include a test that compares the spec that the final installer provides
to the one that it is built with to ensure that they both match exactly.
This should catch any packages that have a mismatching version or any
packages that got added to the installer that don't match the spec.