Check for noarch Python packages
As `constructor` does not correctly work with `noarch: python` packages,
we need to test that they are not included in the installer. Hence we
add a quick check to verify that no `noarch: python` packages were
included in our installer. If any show up, we fail the checks. This
happens even if the user specified them in our specification file.