Do a dry run of `conda update` during the checks
To get an idea of what packages might be out of date, do a dry run
of `conda update` during the checks outputting any information found.
Ideally there should be no changes, but this will provide an idea of
what changes might happen if any. Could also tighten this down later and
make it a failure for an upgrade to happen.